Day 16 (Part 1): Genetic Algorithm
This genetic algorithm runs several iterations, and it can constistently
find the best solution for the example inputs for both parts and the
real input for part 1, albeit it is not especially fast. However, the
search space for part 2 proves too large for this approach to be at all
effective; there are simply too many local maxima. An attempt was made
to overcome this by replacing or combining the algorithm with a
simulated annealing-based approach, but this did not improve the
situation.

Note that, in an attempt to ignore as much of the problem's complexity
as possible, the only rule for solutions generated via inital
generation, crossover, or mutation is that they be "legal": each valve
is reachable from the one preceding it and the one following it. If the
current valve is a neighbor of the last one, this indicates travelling
to it. If it is the same as the last one, this indicates opening it if
is closed and doing nothing otherwise.

In addition to the common "bit flip" mutation (here refered to as a
"replace mutation"), a problem-specific method of mutation ("shift
mutation") was also used. This was done because it was observed that
many optimal patterns were generated at the wrong position within the
solution. Crossovers and replace mutations are ineffective at
horizontally shifting existing patterns, so a shift mutation was
implemented to speed this process up. It noticeably improved the
algorithm's convergence speed.

Many other techniques were also attempted to improve the results of the
genetic algorithm (either its convergence speed or ability to avoid
local maxima). For example, the number of random solutions added into
each generation was steadily increased and reset when a new optimal
fitness was discovered. In another test, older solutions were assigned a
fitness penalty in order to encourage newer/random solutions to overcome
local maxima. These and other techniques did not significantly improve
the algorithm.

In retrospect, genetic and simulated annealing algorithms are not a
very good choice in a situation where the actual optimal solution is
always desired; the only option is running them several times and hoping
to get lucky. Genetic algorithms in particular are designed to find
acceptable local maxima, and it is difficult to avoid this.