Day 16 (Parts 1 & 2): Recursive Search
In light of the mediocre results of the genetic and simulated annealing
algorithms, the solution was rewritten using the Floyd-Warshall
algorithm followed by a recursive DFS in order to exhaustively search
the search space. Some optimizations were made, as explained in the
comments, bringing the total runtime in part 2 down to about 50 minutes.